refactor(general): move tabs to common + wording updates
