--- frontend/src/pages/organization/interfaces/Tabs.interface.ts
@@ -1,5 +1,0 @@
-export interface IPageTab {
-  id: number;
-  name: string;
-  href: string;
-}
